Upgrade to `windows 0.51`
