+++ CONTRIBUTING.md
@@ -35,5 +35,6 @@ dotnet build .metadata
 From the root of the repository:
 
 ```sh
-cargo run -p api_gen
+cargo install --git https://github.com/microsoft/windows-rs --rev f18811c riddle
+riddle --etc bindings.txt
 ```
