+++ Cargo.toml
@@ -12,16 +12,14 @@ include = ["src", "LICENSE"]
 [dependencies]
 anyhow = "1"
 libloading = "0.8"
+# Autogenerated bindings reference `::windows_core` instead of `::windows::core`
+windows-core = "0.51"
 
 [dependencies.windows]
-# TODO: Switch to windows-core when it is published
-version = "0.48"
+version = "0.51"
 features = [
     "implement",
     "Win32_Foundation",
     "Win32_Graphics_Direct3D12",
     "Win32_Graphics_Dxgi_Common",
 ]
-
-[workspace]
-members = ["api_gen"]
