--- api_gen/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "api_gen"
-version = "0.0.0"
-edition = "2021"
-publish = false
-
-[dependencies]
-anyhow = "1"
-windows-bindgen = "0.49"
-windows-metadata = "0.49"
