+++ release.toml
@@ -6,13 +6,6 @@ sign-tag = true
 publish = false
 
 pre-release-replacements = [
-  {file="README.md", search="amd-ext-d3d = .*", replace="{{crate_name}} = \"{{version}}\""},
+  { file = "README.md", search = "amd-ext-d3d = .*", replace = "{{crate_name}} = \"{{version}}\"" },
 ]
 
-# cargo-release only allows using {{version}} in the commit title when creating one
-# commit across all released packages in this workspace (we only release one package
-# though), or by using the same version for all packages.
-# https://github.com/crate-ci/cargo-release/issues/540#issuecomment-1328769105
-# https://github.com/crate-ci/cargo-release/commit/3af94caa4b9bbee010a5cf3f196cc4afffbaf192
-consolidate-commits = false
-shared-version = true
