--- src/Amd/Ext/mod.rs
@@ -1,8 +1,0 @@
-#![allow(
-    non_snake_case,
-    non_upper_case_globals,
-    non_camel_case_types,
-    dead_code,
-    clippy::all
-)]
-pub mod D3D;
